+++ conjur_api/models/general/conjur_connection_info.py
@@ -15,10 +15,11 @@ class ConjurConnectionInfo:
     Used for setting user input data
     """
 
-    def __init__(self, conjur_url: str = None, account: str = None, cert_file: str = None):
+    def __init__(self, conjur_url: str = None, account: str = None, cert_file: str = None, authn_type: int = 0):
         self.conjur_url = conjur_url
         self.conjur_account = account
         self.cert_file = cert_file
+        self.authn_type = authn_type
 
     def __repr__(self) -> str:
         return f"{self.__dict__}"